+++ agent/snmpd.c
@@ -1290,8 +1290,9 @@ receive(void)
 
         DEBUGMSGTL(("snmpd/select", "select( numfds=%d, ..., tvp=%p)\n",
                     numfds, tvp));
-        if(tvp)
-            DEBUGMSGTL(("timer", "tvp %ld.%ld\n", tvp->tv_sec, (long)tvp->tv_usec));
+        if (tvp)
+            DEBUGMSGTL(("timer", "tvp %lld.%ld\n", (long long) tvp->tv_sec,
+                        (long) tvp->tv_usec));
         count = netsnmp_large_fd_set_select(numfds, &readfds, &writefds, &exceptfds,
 				     tvp);
         DEBUGMSGTL(("snmpd/select", "returned, count = %d\n", count));
